Cargo now switches images when being carried, Hero can no longer carry multiple Cargos at once, Cargo now more smoothly moves with Hero

+++ project1/Vehicle.cpp
@@ -10,7 +10,7 @@
 using namespace Gdiplus;
 
 /// Number of pixels wide and tall a tile is.
-const double TileToPixels = 64;
+const int TileToPixels = 64;
 
 /**
  * Constructor
